feat: implement base clients and models for MCP server support
- Added TicketFieldsClient, TicketFormsClient, CustomObjectsClient, UsersClient, TicketCommentsClient

- Added SearchQueryBuilder for dynamic Zendesk search queries

- Added strongly-typed TicketSideload and UserSideload enums

- Expanded Javadocs with Zendesk API deep links and syntax examples

- Implemented MCP emulation integration tests (McpSchemaDiscoverySpec, McpTicketAnalysisSpec)

+++ src/main/java/lol/pbu/z4j/model/TicketField.java
@@ -227,7 +227,7 @@ public class TicketField {
     @Nullable
     @JsonProperty(JSON_PROPERTY_POSITION)
     @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
-    private Integer position;
+    private Long position;
 
     /**
      * The dynamic content placeholder if present, or the <code>description</code> value if not. See <a href=\"/api-reference/ticketing/ticket-management/dynamic_content/\">Dynamic Content</a>
@@ -307,7 +307,7 @@ public class TicketField {
     @Nullable
     @JsonProperty(JSON_PROPERTY_SUB_TYPE_ID)
     @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
-    private Integer subTypeId;
+    private Long subTypeId;
 
     /**
      * Presented for a system ticket field of type \"tickettype\", \"priority\" or \"status\"
